added fixes for the remix starter
